% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armatrix.plots.r
\name{armatrix.cvs}
\alias{armatrix.cvs}
\title{Plot cvs used in armatrix model}
\usage{
armatrix.cvs(x, xlab = "Year", ylab = "CV", col = c("black", "red"), ...)
}
\arguments{
\item{x}{object of class armatrix}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{col}{color vector of length two}

\item{...}{extra arguments to facet_wrap}
}
\description{
Plot cvs used in armatrix model
}
\details{
...
}
