% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armatrix.plots.r
\name{armatrix.effects}
\alias{armatrix.effects}
\title{Plot effect sizes}
\usage{
armatrix.effects(x)
}
\arguments{
\item{x}{object of class armatrix}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{title}{title}
}
\description{
Plot effect sizes
}
\details{
...
}
