% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armatrix.fit.r
\name{armatrix.fit}
\alias{armatrix.fit}
\title{Model to smooth and fill in year*age matrix}
\usage{
armatrix.fit(year, age, x, cv, shrink.cv = 0.5, ...)
}
\arguments{
\item{year}{vector of years}

\item{age}{vector of ages}

\item{x}{vector of values (e.g., weight-at-age)}

\item{cv}{coefficient of variation}

\item{shrink.cv}{numeric (0 = original values,1 = identical to historic average for all years)}

\item{...}{extra arguments to MakeADFun}
}
\value{
list with model output
}
\description{
Model to smooth and fill in year*age matrix
}
\details{
\subsection{General info}{
   \describe{
Models a matrix (e.g., weight-at-age) using an AR1 process over time, age and cohort (separable extension of two densities, 
see see section 6.4 of https://kaskr.github.io/adcomp/_book/Densities.html). 
This will smooth predictions over the three dimensions and fill in gaps (NA values can be used). 

Developed by Noel Cadigan. 
   }
}
 \subsection{specifics}{
   \describe{
   Log(x) = age + year + cohort + age*year + residual
     \item{age}{fixed effect}
     \item{year}{random AR1 effect}
     \item{cohort}{random AR1 effect}
     \item{age*year}{random separable AR1 effect in age and year}
     \item{residual}{based on shrinkage CVs. see shrink.cv}
   }
}
}
\examples{
d <- expand.grid(year=2001:2010,age=1:10)
d$x <- exp(log(d$age)+rnorm(nrow(d),1,0.1))
d$cv <- d$x/10
fit <- do.call(armatrix.fit,as.list(d))
}
