% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armatrix.plots.r
\name{armatrix.pred}
\alias{armatrix.pred}
\title{Plot predicted armatrix model}
\usage{
armatrix.pred(x, xlab = "Year", ylab = "Age", title = "Predicted values", ...)
}
\arguments{
\item{x}{object of class armatrix}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{title}{title}

\item{...}{extra arguments to facet_wrap}
}
\description{
Plot predicted armatrix model
}
\details{
...
}
