% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armatrix.plots.r
\name{armatrix.predobs}
\alias{armatrix.predobs}
\title{Plot predicted and observed armatrix model}
\usage{
armatrix.predobs(
  x,
  xlab = "Year",
  ylab = "Age",
  title = "Predicted vs Observed",
  ...
)
}
\arguments{
\item{x}{object of class armatrix}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{title}{title}

\item{...}{extra arguments to facet_wrap}
}
\description{
Plot predicted and observed armatrix model
}
\details{
...
}
