% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armatrix.plots.r
\name{armatrix.res}
\alias{armatrix.res}
\title{Plot residuals armatrix model}
\usage{
armatrix.res(
  x,
  xlab = "Year",
  ylab = "Age",
  title = "Standardized residuals",
  size = 1.5,
  col = c(`1` = "blue", `-1` = "red")
)
}
\arguments{
\item{x}{object of class armatrix}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{title}{title}

\item{size}{bubble size}

\item{col}{names color vector (positive, zero, negative)}
}
\description{
Plot residuals armatrix model
}
\details{
...
}
