% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armatrix.plots.r
\name{armatrix.res2}
\alias{armatrix.res2}
\title{Plot residuals armatrix model (dotplots)}
\usage{
armatrix.res2(x, ylab = "Standardized residuals")
}
\arguments{
\item{x}{object of class armatrix}

\item{ylab}{ylab}
}
\description{
Plot residuals armatrix model (dotplots)
}
\details{
...
}
