% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armatrix.plots.r
\name{armatrix.waa}
\alias{armatrix.waa}
\title{Plot waa from armatrix model}
\usage{
armatrix.waa(x, xlab = "Year", ylab = "Weight (kg)", title = "", col = "Age")
}
\arguments{
\item{x}{object of class armatrix}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{title}{title}

\item{col}{col lab}
}
\description{
Plot waa from armatrix model
}
\details{
...
}
