% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble.R
\name{bubble}
\alias{bubble}
\title{Bubble plot}
\usage{
bubble(
  x,
  xlab = "Year",
  ylab = "Age",
  scale = 15,
  col = "black",
  alpha = 0.8,
  arrow = 0
)
}
\arguments{
\item{x}{matrix}

\item{xlab}{x label}

\item{ylab}{y label}

\item{scale}{scaler for the bubble size (scale_range)}

\item{col}{color for the bubbles. if 2 colors are given, they indicate positive and negative values}

\item{alpha}{transparency of the bubbles}

\item{arrow}{for cohort strength. see details}

\item{detatils}{Arrows are draw in according to cohort strength (here average value of cohorts). 0 = no arrows, 1 = all arrows and values in between indicate proportions of total}
}
\description{
Bubble plot
}
\examples{
x <-matrix(rnorm(500,0.5,0.2),nrow=10,ncol=50)
dimnames(x) <- list(age=1:10,year=1951:2000)
bubble(x)
}
