% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caa.clean.R
\name{caa.clean}
\alias{caa.clean}
\title{catch-at-age raw: return length frequencies and raw age-length keys}
\usage{
caa.clean(x, plus = NULL)
}
\arguments{
\item{x}{output from caa.all}

\item{plus}{plus group age}
}
\description{
catch-at-age raw: return length frequencies and raw age-length keys
}
\details{
calculates caa from the samples found with caa.raw
}
