% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caa.raw.R
\name{caa.raw}
\alias{caa.raw}
\title{catch-at-age raw: return length frequencies and raw age-length keys}
\usage{
caa.raw(
  catch,
  lf,
  al,
  tresh.al = 2,
  tresh.lf = 2,
  period.unit = c("month", "quarter")
)
}
\arguments{
\item{catch}{data.frame with columns  year,period,region,gear,catch}

\item{lf}{data.frame with columns year,period,region,gear,length,weight.sample,n, sample.id}

\item{al}{data.frame with columns year,period,region,gear,length,age,sample.id}

\item{tresh.al}{threshold of minimum number of samples required for length frequencies}

\item{period.unit}{whether catch and lf are grouped by month or quarter}
}
\description{
catch-at-age raw: return length frequencies and raw age-length keys
}
\details{
diffuses catch into length classes
}
