% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devel-code.R
\name{catch_expr}
\alias{catch_expr}
\alias{make_catch_fn}
\title{Catch conditions}
\usage{
catch_expr(expr, plans)

make_catch_fn(plans)
}
\arguments{
\item{expr}{the expression to be evaluated}

\item{plans}{the plans from make_plans}
}
\description{
To-do: add docs
}
