% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devel-code.R
\name{catchr_opts}
\alias{catchr_opts}
\title{catchr-specific options}
\usage{
catchr_opts(default_plan = NULL, warn_about_terms = NULL,
  bare_if_possible = NULL, drop_empty_conds = NULL)
}
\arguments{
\item{default_plan}{The default plan}

\item{warn_about_terms}{If \code{FALSE}, will not warn about masking special terms}

\item{bare_if_possible}{If \code{TRUE}, and no conditions are collected, will return the result of the evaluated expression as-is, without encompassing named list.}

\item{drop_empty_conds}{If \code{TRUE}, the sublists for conditions that used \code{collect} but didn't collect anything will be dropped from the list. Otherwise, they will appear as empty sublists.}
}
\description{
To-do: add docs
}
