% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort.R
\name{cohort}
\alias{cohort}
\title{Simple cohort strength calculation based on catch-at-age}
\usage{
cohort(x, min = 2)
}
\arguments{
\item{x}{catch-at-age (age x year)}

\item{min}{min number of observations necessary for average}
}
\description{
Simple cohort strength calculation based on catch-at-age
}
\details{
calculates catch-at-age by cohort
}
\examples{
caa <-matrix(rnorm(500,0.5,0.2),nrow=10,ncol=50)
dimnames(caa) <- list(age=1:10,year=1951:2000)
cohort(caa)
}
