% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help-docs.R
\name{collecting-conditions}
\alias{collecting-conditions}
\title{Collect conditions, without halting processes}
\description{
One of the most useful aspects of \code{catchr} is its ability to catch and 'collect' the conditions (e.g., warnings, errors, messages, etc.) raised by an expression without halting/restarting the evaluation of that expression. This can be particularly useful in a number of scenarios:
\itemize{
\item If you are trying to catch the warning messages from code that takes a long time to run, where having to restart the whole process from square one would be too costly.
\item If you want to collect warnings, messages, and errors from code that is running remotely, where these conditions would not be returned with the rest of the results, such as with the \code{\link[future]{future}} package.
\item If you are running lots of code in parallel and want to log all of the conditions within R, such as in a large-scale power simulation, or with packages such \code{\link[purrr:purrr-package]{purrr}}.
}

Using the \code{collect} term lets you do this. Although the exact details/format of the collection process are still being ironed out across versions, the basic premise will remain the same.  When you use \code{collect}, the captured condition will be added to a list of other conditions of that same type. When the expression is done being evaluated, \code{catchr} will return a named list, where "value" is the output of the expression, and the other named elements are sublists with all their collected conditions.
}
\section{Advanced interactions with collected conditions}{


To-do: add docs
}

