% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct.weight.r
\name{correct.weight}
\alias{correct.weight}
\title{species specific correction of weight in function of fish state (round vs headless,...)}
\usage{
correct.weight(x, id, species = c("cod", "mackerel"))
}
\arguments{
\item{x}{value of weight}

\item{id}{state.id (numeric or string; french or english, e.g. Round / Rond)}

\item{species}{species}
}
\description{
species specific correction of weight in function of fish state (round vs headless,...)
}
\details{
conversion of weights
}
