% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.samples.R
\name{do.lookup}
\alias{do.lookup}
\title{do lookup}
\usage{
\method{do}{lookup}(
  catch,
  lf,
  al,
  alg,
  my.levels,
  period.unit,
  min.lf.samples,
  min.al.samples,
  min.al.fish,
  subsample,
  prob.al,
  parallel = FALSE
)
}
\arguments{
\item{catch}{catch data.frame}

\item{lf}{lf}

\item{al}{al}

\item{alg}{al global}

\item{my.levels}{all levels}

\item{period.unit}{month or quater}

\item{parallel}{logical}
}
\description{
do lookup
}
\details{
workhorse function. Does the actual attribution of samples.
}
