% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.grid.df.r
\name{expand.grid.df}
\alias{expand.grid.df}
\title{expand grid for data frames}
\usage{
expand.grid.df(...)
}
\arguments{
\item{...}{data.frames}
}
\description{
expand grid for data frames
}
\details{
https://stackoverflow.com/questions/11693599/alternative-to-expand-grid-for-data-frames
}
