% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-handlers.R
\name{extract_display_string}
\alias{extract_display_string}
\title{Make a string to display from a condition}
\usage{
extract_display_string(cond, cond_name = NA, include_call = T)
}
\arguments{
\item{cond}{A condition to display or turn into a string}

\item{cond_name}{Either the name of the condition you want to display, \code{NA} if you want the condition name to be assigned by default (the first class of the condition), or \code{NULL} if you don't want the condition type displayed at all.}

\item{include_call}{A logical; if \code{FALSE} the call won't be included in the string even if present in the condition.}
}
\value{
A string
}
\description{
Turn a condition into a string comprised of its message, name, and call, in a variety of configurations.
}
