% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.samples.R
\name{find.samples}
\alias{find.samples}
\title{Find samples}
\usage{
find.samples(df, o, y, p, r, g, period.unit)
}
\arguments{
\item{df}{Data.frame in which to find sample}

\item{o}{Option for searching}

\item{y}{Year}

\item{p}{Period}

\item{r}{Region}

\item{g}{Gear}

\item{period.unit}{Whether catch and lf are grouped by "month" (default) or "quarter"}
}
\description{
Find samples
}
\details{
12 steps to find samples.
}
