% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.species.r
\name{find.species}
\alias{find.species}
\title{find species (the code or name used in a certain database)}
\usage{
find.species(species, base = c("ziff", "nafoA", "nafoB"))
}
\arguments{
\item{species}{species name (character string)}

\item{base}{database name}
}
\description{
find species (the code or name used in a certain database)
}
\details{
Given a certain species name (english), find corresponding codes or names in the given database.Makes use of grep.

If species is set to '' than all available species in the database are given.

Databases included:
 \enumerate{
   \item ziff
   \item{nafoA}
   \item{nafoB}
}
}
