% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.speciescode.r
\name{find.speciescode}
\alias{find.speciescode}
\title{find species codes}
\usage{
find.speciescode(species, language = c("en", "fr"))
}
\arguments{
\item{species}{species name (character string)}

\item{language}{en or fr (english by default)}
}
\description{
find species codes
}
\details{
find species codes
}
