% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catching-n-plans.R
\name{first_muffle_restart}
\alias{first_muffle_restart}
\title{Find the first 'mufflable' restart}
\usage{
first_muffle_restart(cond)
}
\arguments{
\item{cond}{A condition}
}
\value{
A restart or \code{NULL} if none can be found.
}
\description{
This function attempts to return the first available \link[base:conditions]{restart} with the string "muffle" in its name. If the condition is an error, it will attempt to find the first restart named "return_error" (used internally in catchr to return a \code{NULL} value). If the condition is an "interrupt", it will attempt to find the first restart named "resume". If no such restarts can be found, it returns \code{NULL}.
}
