% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.lw.R
\name{fit.lw}
\alias{fit.lw}
\title{fit log length-weight relationship by year or period}
\usage{
fit.lw(
  length,
  weight,
  year = 1,
  period = 1,
  tresh = 150,
  period.unit = c("month", "quarter"),
  regression = c("robust", "ols")
)
}
\arguments{
\item{length}{vector of lengths (log scale)}

\item{weight}{vector of weights (log scale)}

\item{year}{vector of years}

\item{period}{vector of periods}

\item{tresh}{minimum number of observations}

\item{regression}{type of regression (ols or rubust)}

\item{interval}{period is month or quarter?}
}
\description{
fit log length-weight relationship by year or period
}
\details{
find lenght weight-relationships for all possible combinations of years and periods. Works with a minimal threshold for individuals needed, and if not attained 1) looks for data in nieghbouring months and 2) uses data for the entire year
returns coefficients, bias correction factor for use log scale , number of observations and the option used to get enough observations
RObuust regression by default (lts). lms and ordinary ls also possible. See MASS package for details
}
