% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.caa.R
\name{get.caa}
\alias{get.caa}
\title{Get catch-at-age results}
\usage{
get.caa(x, plus = Inf)
}
\arguments{
\item{x}{Output from function get.samples().}

\item{plus}{Value of plus group (numeric).}
}
\description{
Get catch-at-age results
}
\details{
This function calculates catch-at-age number (caan), catch-at-age weight (caaw), mean weight-at-age (waa) and mean length-at-age (laa).
}
