% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.cal.R
\name{get.cal}
\alias{get.cal}
\title{Get catch-at-length results}
\usage{
get.cal(x)
}
\arguments{
\item{x}{Output from function get.samples().}
}
\description{
Get catch-at-length results
}
\details{
This function calculates catch-at-length in number (caan) as well as the percentage of each length in the annual catch.
}
