% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.id.R
\name{get.id}
\alias{get.id}
\title{get sample id}
\usage{
get.id(x)
}
\arguments{
\item{x}{vector of strings}
}
\description{
get sample id
}
\details{
Converts a character string to a unique number for use as the sample id. This approach ensure that every id is unique and independent of future data.frame updates (changes, reordering, etc.)
}
