% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.samples.R
\name{get.samples}
\alias{get.samples}
\title{get samples: return length frequencies and age-length keys corresponding to the catch from a period, gear and region}
\usage{
get.samples(
  catch,
  lf,
  al,
  tresh.al = 2,
  tresh.lf = 2,
  period.unit = c("month", "quarter")
)
}
\arguments{
\item{catch}{data.frame with columns  year,period,region,gear,catch}

\item{lf}{data.frame with columns year,period,region,gear,length,weight.sample,n, sample.id}

\item{al}{data.frame with columns year,period,region,gear,length,age,sample.id}

\item{tresh.al}{threshold of minimum number of samples required for length frequencies}

\item{period.unit}{whether catch and lf are grouped by month or quarter}
}
\description{
get samples: return length frequencies and age-length keys corresponding to the catch from a period, gear and region
}
\details{
Attributes length and age data to catch based on the nearest levels (see 12 steps). Used to calculate catch-at-age and length-frequency distributions.

When catch in a certain year is not attributed to a certain region, gear or period, all samples are taken (e.g., catch with no gear will have samples that could be from any gear)
Age-length keys can have missing values. Predictions are made based on a multivariate normal distribution, and are presumed to have a precision of 0,001. 
Steps:
 \enumerate{
   \item to fill
   \item{...}
}
}
