% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catching-n-plans.R
\name{give_newline}
\alias{give_newline}
\title{Make a string end with a newline character}
\usage{
give_newline(s, trim = FALSE)
}
\arguments{
\item{s}{A string.}

\item{trim}{Indicates whether to remove trailing whitespace before adding newline.}
}
\description{
\code{give_newline} will append a line return ('\\n') to the end of a string if
it doesn't already end with one. There is also the option to remove any trailing whitespace before doing so.
}
