% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-n-clean.R
\name{has_handler_args}
\alias{has_handler_args}
\title{Make sure a function can be a handler}
\usage{
has_handler_args(fn)
}
\arguments{
\item{fn}{A function that is a candidate for being a handler}
}
\description{
This makes sure that a given function doesn't \emph{require} more than one argument to be passed into it, and takes in at least one argument (which is what a \link[base:conditions]{handler} needs).
}
