% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.outlier.R
\name{is.outlier}
\alias{is.outlier}
\title{Identify outliers based on the interquantile range}
\usage{
is.outlier(x, coef = 1.5, na.rm = TRUE, ...)
}
\arguments{
\item{x}{vector}

\item{coef}{}
}
\description{
Identify outliers based on the interquantile range
}
\details{
Identifies outliers based on common rule of coef * the interquantile range. The default coefficient is set to 1.5. A coefficient of three is recommended for extremes.  Returns a vector of true or false.
}
