% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{logLik.armatrix}
\alias{logLik.armatrix}
\title{Log likelihood of armatrix object}
\usage{
\method{logLik}{armatrix}(x)
}
\arguments{
\item{x}{output from fit.armatrix}
}
\description{
Log likelihood of armatrix object
}
\details{
...
}
