% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-handlers.R
\name{make_handler}
\alias{make_handler}
\title{Make catchr handler}
\usage{
make_handler(vals, name)
}
\arguments{
\item{vals}{The values of a 'cleaned' \code{catchr} plan}

\item{name}{The type of condition being handled}
}
\description{
To-do: add docs. This won't work for \code{collect} and \code{muffle} for random user-defined stuff. Not really meant for users, but I exported it anyway.
}
