% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.samples.R
\name{fill.multinom}
\alias{fill.multinom}
\title{Fill gaps in age-length keys}
\usage{
\method{fill}{multinom}(df, acol, lcol, id = NULL, smooth = FALSE)
}
\arguments{
\item{df}{Data.frame with columns length and ages, where NA values need to be replaced}

\item{acol}{Vector of ids of age columns}

\item{lcol}{Id of length column (integer)}

\item{id}{Identifier for warning messages}

\item{zero}{Logical. Allow replacement of zero values.}
}
\description{
Fill gaps in age-length keys
}
\details{
Functions that uses the nnet::multinom function to impute age probabilities for lengths that were found in LF samples, but absent of AL samples.
}
