% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.laa}
\alias{plot.laa}
\title{plot.laa}
\usage{
\method{plot}{laa}(x, ci = FALSE)
}
\arguments{
\item{x}{output of get.caa}

\item{ci}{logical (plot with confidence intervals?)}
}
\description{
plot.laa
}
