% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.quality}
\alias{plot.quality}
\title{plot.attribution}
\usage{
\method{plot}{quality}(x, variable = c("option.lengthfreq", "option.agelength"), annual = FALSE)
}
\arguments{
\item{x}{output of get.samples}

\item{annual}{logical}
}
\description{
plot.attribution
}
