% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.waa}
\alias{plot.waa}
\title{plot.waa}
\usage{
\method{plot}{waa}(x, ci = FALSE)
}
\arguments{
\item{x}{output of get.caa}

\item{ci}{logical (plot with confidence intervals?)}
}
\description{
plot.waa
}
