% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotCatch}
\alias{plotCatch}
\title{barplot for catches}
\usage{
plotCatch(data, x, y, fill = NULL)
}
\arguments{
\item{data}{data.frame}

\item{x}{name of columns with x values}

\item{y}{name of columns with y values}

\item{fill}{name of columns with fill values}
}
\description{
barplot for catches
}
