% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotQuality}
\alias{plotQuality}
\title{plotAttribution}
\usage{
plotQuality(
  x,
  variable = c("option.lengthfreq", "option.agelength"),
  annual = FALSE
)
}
\arguments{
\item{x}{output of get.samples}

\item{annual}{logical}
}
\description{
plotAttribution
}
