% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.armatrix}
\alias{print.armatrix}
\title{Print armatrix object}
\usage{
\method{print}{armatrix}(x)
}
\arguments{
\item{x}{output from fit.armatrix}
}
\description{
Print armatrix object
}
\details{
...
}
