% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.bio.r
\name{read.bio}
\alias{read.bio}
\title{Read bio data}
\usage{
read.bio(..., species = c("cod", "mackerel"), language = c("en", "fr"))
}
\arguments{
\item{...}{arguments to read_fwf (e.g., file, progress, skip_empty_rows,n_max)}

\item{species}{see details}

\item{language}{en or fr (english by default)}
}
\description{
Read bio data
}
\details{
read bio data from Peche SAS program. 
currently only mackerel and cod are supported
}
