% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.lf.r
\name{read.lf}
\alias{read.lf}
\title{Reads length-frequency data from dockside monitoring.}
\usage{
read.lf(file, year = NULL, language = "en", ...)
}
\arguments{
\item{file}{.dat file to read. This file is the output of this SAS program: S:/SAS/Peche/autoexec_peche.sas.}

\item{year}{Vector of years to read (ex: 2015:2020). NULL by default, meaning the function will keep data from all available years.}

\item{language}{Language to use for column names, either "en" (default) or "fr".}

\item{...}{Arguments used with read_fwf() (e.g., progress, skip_empty_rows, n_max).}
}
\description{
Reads length-frequency data from dockside monitoring.
}
\details{
Reads length-frequency data from dockside monitoring. Follows structure of file S:/SAS/Peche/Infile_EntFrq.txt.
}
