% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.nafoA.r
\name{read.nafoA}
\alias{read.nafoA}
\title{read nafoA csv file}
\usage{
read.nafoA(file, year = NULL, species = NULL)
}
\arguments{
\item{file}{file name}

\item{year}{vector of years to read (all by default)}

\item{species}{species name(s)}
}
\description{
read nafoA csv file
}
\details{
Read csv file downloaded from from https://www.nafo.int/Data/STATLANT. Data is cleaned so that it is conform with other DFO data (ziff, carbio, length-frequencies)

See find.species to get the species names used in the database
}
