% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.ziff.r
\name{read.ziff}
\alias{read.ziff}
\title{read ZIFF files}
\usage{
read.ziff(sp, path, year = NULL, language = c("fr", "en"))
}
\arguments{
\item{sp}{Numeric species code (ex: 100 for Atlantic cod). Can also be a vector of multiple numeric species codes.}

\item{path}{Directory where files to read are located. Ex: "//dcqcimlna01a/BD_Peches/Ziff/Format CSV/Fichiers de données/".}

\item{year}{Vector of years to read (ex: 2015:2020). NULL by default, meaning the function will keep data from all available years.}

\item{language}{Language to use for column names, either "fr" (default) or "en".}
}
\description{
read ZIFF files
}
\details{
This function reads Zonal Interchange File format (ZIFF) data from .csv files "Version_totale", available from: //dcqcimlna01a/BD_Peches/Ziff/Format CSV/Fichiers de données/.

Notes:
 \enumerate{
   \item{NAFO Divisions and Subareas all capitalised to ensure consistency with other databases}
   \item{year, trimester and month values are from catch dates. If no catch date is provided, the landing date is used instead.}
   \item{Data can be read in both languages (column names, province names) but catch-at-age functions require the English version.}
}
}
