% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devel-code.R
\name{set_default_plan}
\alias{set_default_plan}
\title{Set default catchr plan}
\usage{
set_default_plan(x)
}
\arguments{
\item{x}{The plan to make default plan. The input follows the same rules as \code{\link{make_plans}}.}
}
\description{
To-do: add docs
}
