% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{signal_custom_condition}
\alias{signal_custom_condition}
\title{Signal custom conditions}
\usage{
signal_custom_condition(msg, type = "custom")
}
\arguments{
\item{msg}{The message you want your condition to have.}

\item{type}{Your name for the 'type' of condition you want to raise.}
}
\description{
If you want to signal your own custom conditions (i.e., not messages or warnings),
you can do so with this function. However, most of the time (other than testing code),
you should probably stick to warnings and errors.
}
