% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spay.R
\name{spay}
\alias{spay}
\title{Standardized proportions at year}
\usage{
spay(x)
}
\arguments{
\item{x}{matrix to be standardized (years in columns, ages in rows)}
}
\description{
Standardized proportions at year
}
\details{
Noel's approach
}
