% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spya.R
\name{spya}
\alias{spya}
\title{Standardized proportions at age}
\usage{
spya(x)
}
\arguments{
\item{x}{matrix to be standardized (years in columns, ages in rows)}
}
\description{
Standardized proportions at age
}
\details{
Noel's approach
}
