% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_cod}
\alias{theme_cod}
\title{ggplot theme for cod}
\usage{
theme_cod(base_size = 10, base_line_size = 0.25, base_rect_size = 0.25)
}
\arguments{
\item{base_size}{ggplot base_size}

\item{base_line_size}{ggplot baseline_line_size}

\item{base_rect_siz}{ggplot base_rect_size}
}
\description{
ggplot theme for cod
}
\details{
JOP theme
}
