% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printing.R
\name{print.catchr_compiled_plans}
\alias{print.catchr_compiled_plans}
\alias{summary.catchr_compiled_plans}
\title{View and print 'compiled' catchr plans}
\usage{
\method{print}{catchr_compiled_plans}(x, ..., show_opts = FALSE,
  total_len = getOption("width"), show_full = FALSE)

\method{summary}{catchr_compiled_plans}(object, ...)
}
\arguments{
\item{x}{The "compiled" plans, i.e., from \code{make_plans()}}

\item{\dots}{Currently unused.}

\item{show_opts}{A logical; if \code{TRUE}, prints the catchr options set for the plans.}

\item{total_len}{An integer; sets the total number of characters each line can span before being cut off with "..."}

\item{show_full}{A logical; if \code{TRUE}, will print out the full length of each line.}

\item{object}{The "compiled" plans, i.e., from \code{make_plans()}}
}
\description{
'Compiled' catchr plans returned by \code{\link[=make_plans]{make_plans()}} look very ugly "naked". These functions make plans understandable at a single glance.
}
