% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-n-clean.R
\name{warn_of_specials}
\alias{warn_of_specials}
\title{Check expressions for special terms}
\usage{
warn_of_specials(qs, names_to_check)
}
\arguments{
\item{qs}{A list of quosures.}

\item{names_to_check}{A character vector of reserved special terms to check the expressions for.}
}
\description{
To-do: write doc
}
