% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devel-code.R
\name{make_catch_fn}
\alias{make_catch_fn}
\alias{catch_expr}
\title{Catch conditions}
\usage{
make_catch_fn(plans, opts = NULL)

catch_expr(expr, plans, opts = NULL)
}
\arguments{
\item{plans}{the plans from make_plans}

\item{opts}{opts}

\item{expr}{the expression to be evaluated}
}
\description{
To-do: add docs
}
