% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devel-code.R
\name{catch_expr}
\alias{catch_expr}
\alias{make_catch_fn}
\title{Catch conditions}
\usage{
catch_expr(expr, args, kwargs)

make_catch_fn(args, kwargs)
}
\arguments{
\item{expr}{the expression to be evaluated}

\item{args}{probably will remove}

\item{kwargs}{probably will remove}
}
\description{
To-do: add docs
}
