% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devel-code.R
\name{catchr_DSL}
\alias{catchr_DSL}
\title{The language of \code{catchr}}
\description{
\code{catchr} implements a small but helpful "domain-specific language" (DSL) to make building condition-handling functions simpler to read and type, somewhat like \href{https://tidyeval.tidyverse.org/}{tidyeval}. Essentially, \code{catchr} reserves special 'terms' that mean something different than they do in the rest of R. When given as part of the input for a \code{catchr} plan, these terms will be substituted for special \code{catchr} functions when used to catch conditions.

These special terms can be inputted as strings (e.g., \code{warning = list('collect', 'muffle')}) or as unquoted terms (e.g., \code{warning = c(collect, muffle)})--\code{catchr} converts the unquoted terms to strings internally regardless, but having them unquoted saves keystrokes and can highlight their special meanings for readability.
}
