% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.samples.R
\name{find.samples}
\alias{find.samples}
\title{find samples}
\usage{
find.samples(df, o, d, dn, g, r, y, yn, p, pn)
}
\arguments{
\item{df}{data.frame in which to find sample}

\item{o}{option for searching}

\item{d}{date}

\item{g}{gear}

\item{r}{region}

\item{y}{year}

\item{yn}{neighbouring year}

\item{p}{period}

\item{pn}{neighbouring period}

\item{nd}{neighbouring date}
}
\description{
find samples
}
\details{
12 steps to find samples
}
