% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.samples.R
\name{do.lookup}
\alias{do.lookup}
\alias{find.samples}
\title{do the lookup}
\usage{
\method{do}{lookup}(catch, lf, al, alg, my.levels, mal, parallel = FALSE)

find.samples(df, o, d, dn, g, r, y, yn, p, pn)
}
\arguments{
\item{df}{Data.frame in which to find sample}

\item{o}{Option for searching}

\item{d}{Date.}

\item{g}{Gear.}

\item{r}{Region.}

\item{y}{Year.}

\item{yn}{Neighbouring years.}

\item{p}{Period.}

\item{pn}{Neighbouring periods.}

\item{nd}{Neighbouring dates.}
}
\description{
do the lookup

Find samples
}
\details{
12 steps to find samples.

12 steps to find samples.
}
