% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.samples.R
\name{find.samples}
\alias{find.samples}
\title{Find samples}
\usage{
find.samples(df, o, d, dn, g, r, y, yn, p, pn)
}
\arguments{
\item{df}{Data.frame in which to find sample}

\item{o}{Option for searching}

\item{d}{Date.}

\item{g}{Gear.}

\item{r}{Region.}

\item{y}{Year.}

\item{yn}{Neighbouring years.}

\item{p}{Period.}

\item{pn}{Neighbouring periods.}

\item{nd}{Neighbouring dates.}
}
\description{
Find samples
}
\details{
12 steps to find samples.
}
