% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.caa.R
\name{get.caa}
\alias{get.caa}
\title{get catch-at-age: return length frequencies and raw age-length keys}
\usage{
get.caa(x, plus = NULL)
}
\arguments{
\item{x}{output from get.samples}

\item{plus}{plus group age (numeric)}
}
\description{
get catch-at-age: return length frequencies and raw age-length keys
}
\details{
Once catch of every level is diffused (get.sampled) by length and age class, this function calculates caa.
 So-called scores were added as an exploratory way to gauge the quality of the estimates each years.
}
