% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.samples.R
\name{get.samples}
\alias{get.samples}
\title{Return length frequencies and age-length keys corresponding to the catch from a period, gear and region to use in catch-at-age calculations}
\usage{
get.samples(
  catch,
  lf,
  al,
  tresh.al = 2,
  tresh.lf = 2,
  period.unit = c("month", "quarter")
)
}
\arguments{
\item{catch}{Data.frame with these columns: year, period, region, gear, catch.}

\item{lf}{Data.frame with these columns: year, period, region, gear, length, weight.unit, n, sample.id.}

\item{al}{Data.frame with these columns: year, period, region, gear, length, age, sample.id.}

\item{tresh.al}{Threshold of minimum number of samples required for length frequencies. By default = 2.}

\item{period.unit}{Whether catch and lf are grouped by "month" (default) or "quarter".}
}
\description{
Return length frequencies and age-length keys corresponding to the catch from a period, gear and region to use in catch-at-age calculations
}
\details{
Attributes length-frequencies and age-length keys to a catch level based on a 12 step approach (see steps below). Used to calculate catch-at-age, weight-at-age and length-frequency distributions.

Regions of the catch (e.g., 3P) can be broader then the sample regions (e.g., 3Pn). Sample regions containing the catch region (such as the 3Pn example) will all be used.
When catch in a certain year is not attributed to a certain region, gear or period, that factor is ignored (e.g., catch with no gear will have samples that could be from any gear)
Age-length keys can have missing values. Predictions are made based on a multivariate normal distribution, and are presumed to have a precision of 0,001. 

Steps:
\enumerate{
 \item year, period, region, gear
 \item year, neighbouring period, region, gear
 \item year, period, gear                 
 \item year, neighbouring period, gear
 \item year, neighbouring period
 \item year
 \item neighbouring year, period, gear, region
 \item neighbouring year, neigbouring period, gear, region
 \item neighbouring year, period, gear
 \item neighbouring year, neigbouring period, gear
 \item neighbouring year, gear
 \item neighbouring year
}
}
