% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-handlers.R
\name{is_catchr_plan}
\alias{is_catchr_plan}
\title{Check if list is a catchr plan}
\usage{
is_catchr_plan(x)
}
\arguments{
\item{x}{An object to test}
}
\description{
Currently, this function just checks whether a list was made via
\code{\link[=make_plans]{make_plans()}}. It does so simply by looking at the class of the list. \cr \cr
In the future, catchr plans may become more complicated and this will become a more useful part of the API.
}
