% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devel-code.R
\name{make_plans}
\alias{make_plans}
\title{Make \code{catchr} plans}
\usage{
make_plans(..., default_plan = NULL)
}
\arguments{
\item{\dots}{Named and unnamed arguments for making plans}
}
\description{
Make \code{catchr} plans
}
