% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.bio.r
\name{read.bio}
\alias{read.bio}
\title{Read bio data}
\usage{
read.bio(
  ...,
  year = NULL,
  species = c("cod", "mackerel"),
  language = c("en", "fr")
)
}
\arguments{
\item{...}{arguments to read_fwf (e.g., file, progress, skip_empty_rows,n_max)}

\item{year}{vector of years to read (all by default)}

\item{species}{see details}

\item{language}{en or fr (english by default)}
}
\description{
Read bio data
}
\details{
Read bio data from Peche SAS program. 
Currently only mackerel and cod are supported (To add a species: fill in data_raw/bio_key.csv)
}
