% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.bio.r
\name{read.bio}
\alias{read.bio}
\title{Reads biological data}
\usage{
read.bio(file, year = NULL, sp, language = "en", ...)
}
\arguments{
\item{file}{.dat file to read. This file is the output of this SAS program: S:/SAS/Peche/autoexec_peche.sas.}

\item{year}{Vector of years to read (ex: 2015:2020). NULL by default, meaning the function will keep data from all available years.}

\item{language}{Language to use for column names, either "en" (default) or "fr".}

\item{...}{Optional arguments used with readr::read_fwf() (e.g., progress, skip_empty_rows, n_max).}

\item{species}{Either "cod" or "mackerel".}
}
\description{
Reads biological data
}
\details{
Read biological data from dockside monitoring. 
Currently only mackerel and Atlantic cod are supported (To add a species: fill in data_raw/bio_key.csv)
}
