% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.lf.r
\name{read.lf}
\alias{read.lf}
\title{read length-frequency data (generated by SAS/citrix)}
\usage{
read.lf(..., language = c("en", "fr"))
}
\arguments{
\item{...}{arguments to read_fwf (e.g., file, progress, skip_empty_rows,n_max)}

\item{language}{en or fr (english by default)}
}
\description{
read length-frequency data (generated by SAS/citrix)
}
\details{
reads biological data from commercial samples (SAS/Citrix extractions)
Follows structure of file S:\ SAS \ Peche \ Infile_EntFrq.txt
}
