% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.nafoA.r
\name{read.nafoA}
\alias{read.nafoA}
\title{read nafoA csv file}
\usage{
read.nafoA(file)
}
\arguments{
\item{path}{path to which to download raw files}

\item{year}{vector of years to read (all by default)}

\item{overwrite}{download again? see Details}
}
\description{
read nafoA csv file
}
\details{
Load and clean downloaded csv data from https://www.nafo.int/Data/STATLANT
}
