% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.ziff.r
\name{read.ziff}
\alias{read.ziff}
\title{read ziff files}
\usage{
read.ziff(speciescode, species, year = NULL, language = c("en", "fr"), ...)
}
\arguments{
\item{speciescode}{numeric species code}

\item{year}{vector of years to read}

\item{language}{en or fr}

\item{...}{arguments to list.files (e.g., path)}
}
\description{
read ziff files
}
\details{
Reads ziff files in csv format. The original names of the files are presumed (Version_totale_...).

Available from \\ dcqcimlna01a \ BD_Peches \ Ziff \ Format CSV \ Fichiers de donnees

See find.species to get the species names/codes used in the database.

Notes:
 \enumerate{
   \item sub opanos all capitalised to ensure consistency with other databases
   \item{year / doy / month are from catch, unless info only available for landings }
}
}
