% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.ziff.r
\name{read.ziff}
\alias{read.ziff}
\title{read ziff files}
\usage{
read.ziff(
  speciescode,
  species,
  year = NULL,
  language = c("en", "fr"),
  format = c("csv", "dat"),
  ...
)
}
\arguments{
\item{speciescode}{species code (numeric, for total csv file)}

\item{species}{species name (character string, for species specific dat files)}

\item{year}{vector of years to read (all by default)}

\item{language}{en or fr (english by default)}

\item{format}{data format (old dat or new csv)}

\item{...}{arguments to list.files (e.g., path)}
}
\description{
read ziff files
}
\details{
Two ziff formats are supported; \cr
old dat format: z_180_....espece.dat -> S:/Ziff/POISSON/DONNEES/ \cr
new csv format: Version_totale_....csv -> \\ dcqcimlna01a \ BD_Peches \ Ziff \ Format CSV \ Fichiers de données
}
