% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devel-code.R
\docType{data}
\name{special_terms}
\alias{special_terms}
\title{To-do:}
\format{An object of class \code{character} of length 9.}
\usage{
special_terms
}
\description{
CODE-BASED:
\itemize{
\item let code use rlang lambda functions for functions
\item figure out how you can avoid overlap between "warning" and "condition"
\item make the towarning, etc. remove the call
\item make a 'misc' condition thing which gets everything that isn't already being caught
\item handle interrupts
\item make the order matter
}
}
\details{
LESS CODE-BASED:
\itemize{
\item add examples
\item make the spec_terms thing use package defaults somehow
\item make a way of stopping warnings from catchr
\item standardize terminology
\item add a page about all the options
\item fill out a page about collecting
\item the biggest thing preventing me from removing the rlang/purrr dependencies is the splicing operator, it seems
\item make a help page that describes how things are masked:
\itemize{
\item The only thing that is masked in evaluation is the non-function versions of the special names.
\item Egh, just make a help page and connect it to the warning message
}
}
}
\keyword{datasets}
